requirements.txt
requests>=2.27,<3
pandas==1.3.5
beautifulsoup4==4.9.3
lxml==4.9.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
tqdm==4.65.0
streamlit==1.28.0
transformers==4.28.1
torch==2.0.1
huggingface-hub==0.25.0  
